changed the data type of inputs

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_income = int(input('Enter your monthly income: '))
-total_monthly_expens = int(input('Enter your total monthly expenses: '))
+monthly_income = float(input('Enter your monthly income: '))
+total_monthly_expens = float(input('Enter your total monthly expenses: '))
 
 monthly_savings = monthly_income - total_monthly_expens
